# front/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Front Service
    description: An HTTP service for accessing the backend services of a multiplayer game
    version: ""
host: localhost:8000
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
    /api/character:
        get:
            tags:
                - Front
            summary: listCharacters Front
            description: List all characters
            operationId: Front#listCharacters
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/CharacterResponse'
            schemes:
                - http
        post:
            tags:
                - Front
            summary: createCharacter Front
            description: Create a new character
            operationId: Front#createCharacter
            parameters:
                - name: CreateCharacterRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FrontCreateCharacterRequestBody'
                    required:
                        - name
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/FrontCreateCharacterResponseBody'
                        required:
                            - name
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/FrontCreateCharacterCharacterAlreadyExistsResponseBody'
                        required:
                            - message
                            - name
            schemes:
                - http
    /api/character/{name}:
        get:
            tags:
                - Front
            summary: getCharacter Front
            description: Get a character by name
            operationId: Front#getCharacter
            parameters:
                - name: name
                  in: path
                  description: name of character to be retrieved
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/FrontGetCharacterResponseBody'
                        required:
                            - name
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/FrontGetCharacterCharacterNotFoundResponseBody'
                        required:
                            - message
                            - name
            schemes:
                - http
        put:
            tags:
                - Front
            summary: updateCharacter Front
            description: Update a given character
            operationId: Front#updateCharacter
            parameters:
                - name: name
                  in: path
                  description: Name, which is a unique identifier
                  required: true
                  type: string
                  maxLength: 32
                - name: UpdateCharacterRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FrontUpdateCharacterRequestBody'
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/FrontUpdateCharacterResponseBody'
                        required:
                            - name
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/FrontUpdateCharacterCharacterNotFoundResponseBody'
                        required:
                            - message
                            - name
            schemes:
                - http
        delete:
            tags:
                - Front
            summary: deleteCharacter Front
            description: Delete a given character
            operationId: Front#deleteCharacter
            parameters:
                - name: name
                  in: path
                  description: name of character to be deleted
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
            schemes:
                - http
    /api/inventory/{character}:
        get:
            tags:
                - Front
            summary: getInventory Front
            description: Get a character's inventory
            operationId: Front#getInventory
            parameters:
                - name: character
                  in: path
                  description: name of character
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            type: string
                            example: Distinctio nihil.
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/FrontGetInventoryCharacterNotFoundResponseBody'
                        required:
                            - message
                            - name
            schemes:
                - http
    /api/inventory/{character}/{item}:
        post:
            tags:
                - Front
            summary: addInventoryItem Front
            description: Add an item to a character's inventory
            operationId: Front#addInventoryItem
            parameters:
                - name: character
                  in: path
                  description: character's name
                  required: true
                  type: string
                - name: item
                  in: path
                  description: item's name
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/FrontAddInventoryItemItemNotFoundResponseBody'
                        required:
                            - message
                            - name
            schemes:
                - http
        delete:
            tags:
                - Front
            summary: removeInventoryItem Front
            description: Remove an item from a character's inventory
            operationId: Front#removeInventoryItem
            parameters:
                - name: character
                  in: path
                  description: character's name
                  required: true
                  type: string
                - name: item
                  in: path
                  description: item's name
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/FrontRemoveInventoryItemCharacterNotFoundResponseBody'
                        required:
                            - message
                            - name
            schemes:
                - http
    /api/item:
        get:
            tags:
                - Front
            summary: listItems Front
            description: List all items
            operationId: Front#listItems
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/ItemResponse'
            schemes:
                - http
        post:
            tags:
                - Front
            summary: createItem Front
            description: Create a new item
            operationId: Front#createItem
            parameters:
                - name: CreateItemRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FrontCreateItemRequestBody'
                    required:
                        - name
                        - description
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/FrontCreateItemResponseBody'
                        required:
                            - name
                            - description
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/FrontCreateItemItemAlreadyExistsResponseBody'
                        required:
                            - message
                            - name
            schemes:
                - http
...
definitions:
    CharacterResponse:
        title: CharacterResponse
        type: object
        properties:
            description:
                type: string
                description: Description
                default: ""
                example: Savior of Princess Zelda
                maxLength: 128
            experience:
                type: integer
                description: Amount of experience
                default: 0
                example: 2285596
                minimum: 0
                maximum: 1e+07
            health:
                type: integer
                description: Current health
                default: 100
                example: 2650467
                minimum: 0
                maximum: 1e+07
            name:
                type: string
                description: Name, which is a unique identifier
                example: Link
                maxLength: 32
        description: Attributes of the character
        example:
            description: Savior of Princess Zelda
            experience: 470847
            health: 7549643
            name: Link
        required:
            - name
...
    FrontCreateCharacterRequestBody:
        title: FrontCreateCharacterRequestBody
        type: object
        properties:
            description:
                type: string
                description: Description
                default: ""
                example: Savior of Princess Zelda
                maxLength: 128
            experience:
                type: integer
                description: Amount of experience
                default: 0
                example: 2471470
                minimum: 0
                maximum: 1e+07
            health:
                type: integer
                description: Current health
                default: 100
                example: 5111253
                minimum: 0
                maximum: 1e+07
            name:
                type: string
                description: Name, which is a unique identifier
                example: Link
                maxLength: 32
        example:
            description: Savior of Princess Zelda
            experience: 9877087
            health: 2753424
            name: Link
        required:
            - name
...